V2.1
- Improved startup_gd32fxxx.S
- CMSIS has its own project now
- Upgraded to CMSIS V5.0.5
- Removed lib-debug. Now part of lib-hal
- Updated to latest GD32 Firmware
- Introduced debug software UART0
- Major cleanup lib-configstore
- In order to support Newlib; removed lib-c, lib-c++ and added lib-clib
- In order to support FreeRTOS; Added Software timers in H.A.L. framework.
- Major update network stack. Added support for RFC 5227 and RFC 3927
- Updated Art-Net with latest specifications
- Improved support for RDMNet LLRP only
- Moved parameter utc_offset to "env.txt"
- Improved HTTP Server
- Added support for remote UART shell
- Removed lib-rdmdiscovery, lib-rdmnet and lib-rdmresponder
- Removed lib-oscclient and lib-oscserver

+++ lib-osc/src/oscsimplesend.cpp
@@ -2,7 +2,7 @@
  * @file oscsimplesend.cpp
  *
  */
-/* Copyright (C) 2020-2023 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2020-2024 by Arjan van Vught mailto:info@gd32-dmx.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -122,5 +122,5 @@ void OscSimpleSend::UpdateMessage(const char *pPath, uint32_t nPathLength, char 
 void OscSimpleSend::Send(uint32_t nMessageLength, int32_t nHandle, uint32_t nIpAddress, uint16_t nPort) {
 	debug_dump(s_Message, nMessageLength);
 
-	Network::Get()->SendTo(nHandle, s_Message, static_cast<uint16_t>(nMessageLength), nIpAddress, nPort);
+	Network::Get()->SendTo(nHandle, s_Message, nMessageLength, nIpAddress, nPort);
 }
